+++ 3. Study JPA/SpringBootOnTodo/src/main/java/edu/doubler/todo/repository/UserRepository.java
@@ -6,4 +6,4 @@ import org.springframework.stereotype.Repository;
 import edu.doubler.todo.entity.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long>{}
+public interface UserRepository extends JpaRepository<User, Integer>{}
